move rfc2047 functions to library
