+++ mutt/mutt.h
@@ -70,6 +70,7 @@
 #include "message.h"
 #include "mime.h"
 #include "regex3.h"
+#include "rfc2047.h"
 #include "sha1.h"
 #include "signal2.h"
 #include "string2.h"
